fix(api): scope members middleware per route; pin the no-workspace accept path
Review found a merge-blocker: members mounted a `.use("*")` middleware, and a
Hono sub-app's wildcard middleware is registered at the MOUNT PREFIX — so it
also ran for sibling sub-apps mounted at /api afterwards, putting
requireWorkspace in front of /api/invites/accept. A brand-new invitee has no
workspace and cannot send x-workspace-id, so the whole token-bearer flow 400'd
in the real app. The e2e suite hid it by composing invites-first.

- members.ts: attach requireSession/requireWorkspace per route, so the mount
  order in index.ts cannot affect any sibling router
- invites.e2e: compose workspaces → members → invites, the SAME order as
  index.ts, and add the regression pin (preview + accept with no
  x-workspace-id header at all). Restoring `.use("*")` fails 13 tests.
- move isUniqueViolation out of lib/messages.ts into lib/db-errors.ts: it is a
  generic drizzle-error walker, and the invites path importing it from the
  message writer put an unrelated consumer on that module (standards review)
